### 2019-08-12 #1645 @danielskatz
👋 @RHagenson  - please add to your paper an explantation of what research problems your software can help with (which is another way of saying who will cite your software).  Perhaps add a little about how they would use your software.  When you have done so, please enter ```@whedon generate pdf``` as a new comment here to regenerate the pdf.
