### 2019-08-16 #1645 @lpantano
Hi @RHagenson, thanks for submitting to JOSS. I look around but I couldn't find any documentation of the tool. One of the mandatory requirement is to have full documentation with examples. If it is already done, can you put the link in the readme of the repo?

Cheers
