### 2019-08-16 #1645 @RHagenson
@lpantano Documentation can be built through `godoc`, for online access to documentation visit: https://godoc.org/github.com/bio-ext/bio-go/bio

I have been using property tests to serve as both tests and examples. These can be run and seen from the repo root through `go test` via `go test -v ./...` (I have also just now added the `-v` flag to produce this output in TravisCI). In my opinion, code that is not automatically validated with each push to `master` is less useful than strict properties that can be expected from the library. The intent of the property tests (perhaps not as well as I had hoped) is to show potential users what properties they can expect from an implementation. All this said, I agree that explicit examples have value so I will add an `examples/` directory alongside `bio/` at the root. The separate directory would allow running everything all at once in CI and let potential users clearly see where examples are housed without all the property-testing framework around them.

I will add all the above directive information on viewing documentation and running tests to the README later today.

In your opinion, are there other additions I need to make?
