### 2019-08-16 #1645 @lpantano
Thank you for the detailed explanation, I think that would help a lot to show potential users how they can use your tool, and then the goDoc can get into details.

Thanks for making the changes.
