### 2019-08-17 #1645 @RHagenson
@lpantano I added the documentation and testing sections to the README. However, due to my immediate schedule I will not be able to add `examples/` quite yet. I will tag you again in this thread when I have `examples/` added. If there are any other additions/changes I need to make, let me know.
