### 2019-08-20 #1645 @RHagenson
@lpantano Not done with the examples (they are being developed on `examples` branch), but I wanted to update you that I will be doing two types of examples:
1. Individual documentation example, i.e. examples that will display in the documentation built with `godoc` that show usage of individual elements
2. Aggregate examples, i.e. examples that use multiple elements together
