### 2019-08-22 #1645 @peterjc
I was asked (by email) to review, but declined:

Sadly I know nothing about Go, and the project's current state of documentation leaves too steep a learning curve for me or any other bioinformatician curious to try it.

It needs a quick start, I would think even a trivial example like loading a given DNA sequence (be it with IUPAC ambiguity codes or just A, C, G, T bases) from a FASTA file, translating it into an amino acid sequence, and then writing out this protein sequence as a new FASTA file would be useful. Looking at the module names I think this functionality would be covered already.
