### 2019-08-22 #1645 @RHagenson
@peterjc Thank you for the advice on a helpful and explanatory aggregate example of usage reading in a given DNA FASTA, translating it to protein, and writing it out as a new FASTA. In your opinion, are there any other rough edges about the project in its current state that need smoothing?
