### 2019-08-22 #1645 @peterjc
@RHagenson the beginner documentation was the main thing lacking - beyond that it was hard to say more, sorry.
