### 2019-08-26 bio-ext/bio-go#1 @RHagenson
Add docs/ on master (GitHub Pages)
To aid new users, there needs to be a rendered webpage detailing usage and best practices of `bio-go` aimed at new Go users that complements the more intermediate/advanced source documentation built via `godoc`. Rather than a `gh-pages` branch, I think it best to have the new user documentation exist inline with the rest of the code on `master`.

One of the first examples/writings should be a reading in (DNA) FASTA, translating it, and outputting the (Protein) FASTA -- complements to https://github.com/openjournals/joss-reviews/issues/1645#issuecomment-523822852.
