### 2019-08-28 #1645 @will-rowe
Happy to review this too if needed
