### 2019-08-30 bio-ext/bio-go#1 @RHagenson
Decision made to use the Just the Docs theme (https://jekyllthemes.io/theme/just-the-docs) as it has a clear design intended to focus writing and reading on critical information (and has a decent set of bells and whistles to it that could be used later).
