### 2019-09-04 bio-ext/bio-go#1 @RHagenson
Work has begun in dedicated dedicated `docs` branch, see d89365c
