### 2019-09-10 bio-ext/bio-go#1 @RHagenson
Initially merged into master on 0b6c1155fb94e722eeee181ca153f7e2dce564f2

Still more work to be done by providing more details of each of the subpackages thus far. This plaintext/non-code documentation of packages should be housed under its own major/parent category the same as Examples are housed currently.
