### 2019-09-16 #1645 @RHagenson
@lpantano, I finished writing the beginner documentation (available at https://bio-ext.github.io/bio-go/) and the updated version number is [v1.1.3](https://github.com/bio-ext/bio-go/tree/v1.1.3).
