### 2019-09-17 #1741 @whedon
[REVIEW]: bio-go: A flexible tools Bioinformatics library
**Submitting author:** @RHagenson (<a href="http://orcid.org/0000-0001-9750-1925">Ryan Hagenson</a>)
**Repository:** <a href="https://github.com/bio-ext/bio-go" target ="_blank">https://github.com/bio-ext/bio-go</a>
**Version:** v1.0.0
