### 2019-09-17 #1741 @RHagenson
Note the version number under review should be [v1.1.3](https://github.com/bio-ext/bio-go/tree/v1.1.3) as v1.0.0 (as stated above) was prior to implementing pre-review comments.
