### 2019-09-18 bio-ext/bio-go#5 @will-rowe
JOSS review
Hi,

Thank you for this contribution to the bioinformatics community! I am filing this issue as part of my review for [your submission to JOSS](https://github.com/openjournals/joss-reviews/issues/1741). Please could you address or respond to these comments.

Best,

Will

## general

On the whole, this is a good contribution which has some nice features, clear design and is well written. However, there are a few issues that need addressing and I detail these in the sections below.

As a more general comment, I was a bit disappointed in the choice of name for this repository, and to a lesser extent the library name. As there is already a well established bioinformatics library for Go named `biogo`, another similarly named library is confusing; particularly when taking into account the full package import path (`github.com/bio-ext/bio-go/bio/<package>`). The documentation site also refers to the project as `bio-go`, wheras the github repository refers to it as `bio`.

Aside from being a bit confusing, this may reduce the impact of your contribution by potential users not realising `bio` is a new alternative to `biogo`! I don’t expect you to change the name, I just wanted to raise the point.

## code and repository

* You can't use `@` in `bio/io/fastq` as a record separator. FASTQ quality scores also utilise the `@` character, so any sequence read with an initial base quality of `@` will break the record parser.

* As stated in your open issues, a guideline for contributing is needed.

* Please consolidate the library naming throughout the `repository/godoc/documentation`.

## paper

* I think you need to be more thorough in your review of the field. There is a checkbox on the review criteria that asks if the author has compared their contribution and I don't feel you have done this yet.

* It would be nice to acknowledge or cite [biogo](https://joss.theoj.org/papers/10.21105/joss.00168) in the paper. It would definitely benefit users to know what advantages and disadvantages `bio` has over `biogo`.

* I found your explanation of the design structure much clearer in the repo readme, compared to the explanation in the paper. Maybe consider updating the explanation within the paper.

## documentation

* As it is a review checkbox for the documentation, I think you should copy the explicit `statement of need` from your paper into your documentation, probably where you also have your `philosophy` and `approach` statements.

* The example is good but it would be clearer if you gave a one line summary at the top of the page, or have a more intuitive title. So rather than just calling it `central dogma`, state that the example is going to create a program to convert DNA sequence to protein sequence.

* I would stick to calling it a `Command Line Interface` rather than interface, as this can be confused with Go’s interfaces. So, instead of `set up the interface` as a header, adjust to `set up the Command Line Interface (CLI)` or similar.

* In the `Approach` section of the documentation you make the case that the package import path aids code readability, yet you use package aliasing in the example (importing the `fasta/base` as `fa`). Although you explain why you have used aliasing in the example, for the sake of continuity and readability, I would avoid the aliasing in the example, or remove the emphasis you placed on this in the previous `Approach` section.

* Some of your variable names are uninformative, which I think is unhelpful in an example. I think readability would be improved by changing things like the `x` in `x, err := os.Open(*input)` to something like `fileHandle` or just `file`

* The `fun go fact` / `performance optimisations` in the packages sections is fairly repetitive. It's a nice idea but in my opinion is currently getting in the way. Maybe you could have a separate page and group the facts as performance opimisations?

* It feels as though some pages in the packages section are incomplete. For instance, in `io/fasta` you mention the reader in a footnote but this isn't included anywhere else on the page.

* Broken link to navigate back to `io` from `io/fasta` or `io/fastq`.

* I feel like another example, or a more detailed example, would really strengthen the documentation (and library in general). It feels a little thin at the moment.
