### 2019-09-18 #1741 @will-rowe
Hi @RHagenson!

I've just done my review and have filed it as an issue over at the [repo](https://github.com/bio-ext/bio-go/issues/5). Please let me know if you need anything clarifying.

Cheers,

Will
