### 2019-09-18 bio-ext/bio-go#5 @RHagenson
@will-rowe Thank you for your thorough review of `bio-go`. I will begin addressing all your concerns later today.
