### 2019-09-19 bio-ext/bio-go#5 @RHagenson
@will-rowe as you can tell from the series of "This was referenced X hours ago" statements above, I split your comments into new issues so I can better ensure I address each point. You bring up some really strong points of concern and identified weaknesses of `bio-go` that I had not considered.

Bit of brainstorming on my end here for the unfortunate naming clash of `bio-go` and `biogo`, which I agree is problematic -- point really driven home as I wrote out #8. Since `bio-go` exists as a "Semantic Bioinformatics for Go" that exposes equitable solutions, how problematic do you see the name `sembio-go` to evoke the shortening often given to Semantic Versioning as "semver" while also remaining clear that it is a repo for **bio**informatics in Go? If a name change is in order (which I increasingly believe it should be done), now, before any active users, is the time to do so.
