### 2019-09-19 bio-ext/bio-go#19 @kortschak
JOSS review issues
@RHagenson this is a collection of  issues to address for the JOSS review process.

Would you please address these so I can check off the criteria at openjournals/joss-reviews#1741.

- [ ] There are some functions that are not currently covered by tests. While a couple are trivial, the [`fastq.Write`](https://codecov.io/gh/bio-ext/bio-go/src/32c04156a1f85308e881cd81af0d17760e49877c/bio/io/fastq/write.go) is completely uncovered.
- [ ] I don't believe that the FASTQ reading code is robust to malformed input. This should be tested for.
- [ ] The FASTA reading code is not tested. This is essential.
- [ ] For the I/O routines, I would strongly recommend using the go-fuzz tool to ensure correctness. This is not critical for the review, but would be worthwhile.

- [ ] There should be a statement of need either in the repository's readme or the top level godoc (or both).
- [ ] There is good documentation for the leaf packages, but the intermediate interfaces could probably do with examples given the stated goal of provided an approach for solution space searching by bioinformaticians.
- [ ] Please add community guidelines as per the JOSS standard.
- [ ] I am curious about the utility of the `utils.StringsToBytes` and `utils.BytesToStrings`. Without an example, it is difficult to know what these would actually be used for. Also note typos in `BytesToStrings` doc comment.
- [ ] Please reconsider the naming of the utils package. This is a significantly non-Go-idiomatic package name.
- [ ] `tests.DeepClone` should be marked as unsafe for use with cyclic data structures; this is a consequence of its use of gob as a backing implementation.
- [ ] If tests are for use only within the bio-ext packages (which it appears to be), I would recommend placing it in an internal package tree.

- [ ] It would be good to know how the packages here interface with other packages already existing in the space. This would not be limited to bíogo, but also possibly the Gonum scientific software suite.
