### 2019-09-19 bio-ext/bio-go#19 @RHagenson
@kortschak, thank you for your review. I will begin addressing your concerns.
