### 2019-10-30 #1741 @RHagenson
@danielskatz Pausing it is reasonable. I will continuing making the major changes then resume review once they are implemented. Thank you for your understanding.
