### 2020-02-03 #1741 @rhagenson
@lpantano, I wanted to inform you that as part of the review process where both reviewers suggested a name change I have changed the organization and repo name. The code now resides at: https://github.com/sembio/go and the paper title has changed accordingly to "sembio/go: Semantic Bioinformatics for Go" -- I did not change significant portions of the code during this migration.

I still need to update the Zenodo archival process and make the other corrections as reviewers noted. Review should remain paused until I have made the full migration to the new code repo.
