### 2020-03-10 #1741 @rhagenson
@danielskatz Thank you for bumping this to the top of my list.

Unfortunately, through this review process I have realized that `bio-go` (now `sembio/go`) is not the priority it should be for me to maintain a project of its scale in the longterm. I still believe `sembio/go`'s approach is valuable. However, I do not currently have the necessary bandwidth to maintain, extend, and improve `sembio/go`.

With this said, I wish to regretfully withdraw this submission in its current state. Thank you to all who have committed time to reviewing and managing this submission.
